Implement Archive Repacking and Extract -k flag
This is a big commit with several bad choices in it, that will need to
be fixed. But it works! And it works pretty well.

This can rebuild a bit-for-bit identical copy of the `audio` archive.
Due to the folder markers, it can't rebuild `dd` accurately, but the
game will still take our rebuilt archive.